Addressing code comments.
