+++ client/package.json
@@ -13,7 +13,7 @@
     "react-dom": "^18.2.0",
     "react-router": "^6.3.0",
     "react-router-dom": "^6.3.0",
-    "react-scripts": "^2.1.3",
+    "react-scripts": "5.0.1",
     "w3-css": "^4.1.0",
     "web-vitals": "^2.1.4"
   },
